requirements.txt
Django==2.2
mysqlclient==2.0.0
djangorestframework==3.10.3
django-cors-headers==3.2.0
gunicorn==20.0.4
websockify==0.9.0
supervisor==4.1.0
